all: run tests on postgres v10 (#2413)
### What
Run tests on Postgres v10 and add function to `support/db/dbtest` that exposes the current version of Postgres that's running.

### Why
Recoverysigner (`exp/services/recoverysigner`) is going to be using the IDENTITY feature of Postgres that was introduced in v10 which makes it easier and safer to copy records into an audit table, as well as being the recommended way of doing auto-incrementing primary keys now. Recoverysigner's tests are going to use the new version function to skip themselves if the tests are being run against an older version of Postgres.

Also I heard that at some point we'll be wanting to support Postgres v11 on Horizon and run tests for it, and so now seems like a good time to start running tests for a newer version of Postgres against to provide some confidence towards that in the future.

### Known limitations

This will increase build times a little because more builds will need to finish.

### Alternative approaches considered

We could use Go build tags so that the tests that require Postgres 10 are separated and only run in certain cases, but it's really annoying on developers to need to add build tags when running Go commands and it will be easy to accidentally not run the tests. Skipping is much simpler for a developer to interact with. Tests that are skipped will output a special message saying why they were skipped.