exp/services/recoverysigner: remove MemoryStore (#2428)
### What
Remove the MemoryStore that stores accounts in memory.

### Why
We don't need it anymore. It existed temporarily until we had the Postgres integration complete. That is complete now and all tests and deployments are using Postgres.

As part of the change the `--db-url`/`DB_URL` config option no longer defaults to an empty string. The empty string indicated the memory store should be used, but since the memory store is gone there is little value in retaining the empty string default and a typical localhost default makes it easier to run the app locally when developing. This is the default we use in the migration command too.