services/horizon/internal: Only use repeatable read transactions on state endpoints (#2454)
In #2416 we extended the Last-Ledger header to cover additional endpoints which were not immutable.

@2opremio raised the following point before working on the PR:

> After giving this quite a bit of thought, isn't it overkill to set that header (and in turn access the DB for all endpoints) to make horizon-cmp happy?

After testing the 1.1.0 release candidate on staging we found evidence to confirm this hypothesis. The overhead of wrapping each request in a repeatable read transaction proved to significantly hinder performance.

We should remove this change from the 1.1.0 release. We can always reintroduce this change later on but at this point it would be better to explore other solutions to the underlying problem we are trying to solve (verifying correctness of a horizon release) that have less of a negative impact on performance.