services/horizon/internal/db2/history: Fix ingestion of max_fee and fee_charged transaction fields (#2456)
We changed the representation of thefee_charged column in postgres from integer to bigint in https://github.com/stellar/go/pull/2429/files . However, I forgot to update the code which inserted transaction rows to cast the fee_charged field as an int64 instead of an int32.

The XDR representation of the MaxFee attribute of a v1 and v0 transaction envelope is an unsigned 32 bit integer. We have always represented the max_fee column in postgres as an integer type. Postgres can only store signed integers. This means the maximum value a postgres integer can hold is 2^31 - 1 = 2147483647. So if you try to insert an unsigned 32 bit integer which is larger than 2^31 -1, e.g. 2500000000, you will encounter this error: pq: value \"2650000000\" is out of range for type integer.

There are two ways to fix this:

(1) cast max_fee to a 32 bit signed integer when inserting transactions into the db. When you read transactions from the db you have to remember to cast the 32 bit signed integer into an unsigned 32 bit integer.
(2) change the column type of max_fee to bigint

The reason this postgres error didn't manifest before is because, previously, we were casting max_fee to a 32 bit signed integer:

https://github.com/stellar/go/blob/horizon-v1.0.1/services/horizon/internal/db2/history/transaction_batch_insert_builder.go#L156

However, we did not cast the 32 bit signed integer back into an unsigned 32 bit integer when reading rows from the db. Consequently, the "fee_charged" field in the Horizon response is negative for some transactions.

I have opted for changing the column type of max_fee to bigint because it is a more robust solution. We do not have to have any special decoding logic when reading transactions out from the db.

Although changing the column type of max_fee requires rebuilding the history_transactions table, we are already doing that work for the fee_charged. Adding an additional clause to the ALTER TABLE statement should not lengthen the time required to run the migration.