protocols/horizon: Handle MaxFee and FeeCharged as ints or strings when parsing JSON (#2492)
In Horizon v1.1.0 we changed the type of FeeCharged and MaxFee from int32 to int64. Once Horizon SDKs have been prepared for the type change, Horizon will need to start encoding those fields as strings when serializing to JSON. That is necessary because JSON and javascript do not support 64 bit numbers as a native type.

When the Horizon Go SDK unmarshalls the Horizon Transaction JSON it should be able to handle responses which have the fields encoded as ints and responses which have the fields encoded as strings.