exp/services/recoverysigner: add Firebase setup instructions (#2496)
### What

Add Firebase setup instructions for setting up a Firebase project for use with recoverysigner.

### Why

Recoverysigner currently uses Firebase as an authentication method for email address and phone number but it may not be clear to someone using recoverysigner how to setup the Firebase project to work with recoverysigner. While this doesn't replace having expert knowledge and experience of Firebase or Firebase's own documentation this should be enough to point a developer in the right direction for what needs configuring.