all: Update txnbuild to support fee bump transactions (#2495)
This PR refactors txnbuild.Transaction to support features introduced in the upcoming protocol 13 release (in particular CAP 15 and CAP 18). The new API for txnbuild is not backwards compatible which is why the version for this release is a major version bump from the last horizon client release.

Users of txnbuild and horizonclient should upgrade to v3.0.0 before protocol 13 is enabled. Pre v3 clients will still be able to submit transactions to a protocol 13 enabled stellar network. However, pre v3 clients will not able to parse xdr strings corresponding to protocol 13 transactions.

It is safe to upgrade to v3.0.0 before protocol 13 is enabled because we have configured txnbuild to construct transactions which are compatible with both protocol 12 and protocol 13. After protocol 13 is enabled, we will create a v3.1.0 release of horizon client which configures txnbuild to construct transactions which are only compatible with protocol 13.