exp/services/recoverysigner: add basic logs (#2567)
### What

Add basic logs for each action the service can perform that describe what's happening to help with identifying issues. Also fixes logs so that all existing logs carry the request ID and account ID for the current request.

### Why

This is a young service and we're likely to run into issues that we will need logs to help us debug.