support/http: log user agent header at request start (#2619)
### What
Log the `User-Agent` header at request start.

### Why
Identifying which agent sent a request can be useful in the reference
implementations that accept inbound requests. It can be useful for
debugging strange issues, understanding broad usage patterns, and if a
service is deployed in use with a specific client and that client uses
the User-Agent header to communicate its version it can help an operator
understand what versions of its client are accessing the service. This
can help make decisions about upgrade plans, etc.

The specific use case I have is that we are deploying webauth and
recoverysigner with a client and I want to be able to know which version
of the client is accessing these services.

Horizon and the Horizon SDKs actually use some custom headers to achieve
this specific goal but I am not repeating that pattern since it is more
common to communicate this information in the user agent header.

The header does not normally contain any personally identifiable
information or sensitive information and so I believe there are no
concerns on that front.

This change impacts these services inside this repo:
 - `exp/services/webauth`
 - `exp/services/recoverysigner`
 - `services/friendbot`
 - `services/federation`