exp/service/recoverysigner: add signers to response for SEP-30 v0.3.0 (#2614)
### What
Add a list of all signers that the server has available for an account consistent with the new field as added in SEP-30 v0.3.0 stellar/stellar-protocol#619.

### Why
SEP-30 v0.3.0 adds this field to support signer rotation. The server can generate new signers at anytime and clients can choose to switch out their old signer on their account to the new signer at a time that is convenient for them.

This implementation of the server still only supports a single key but it supports the endpoints that will support the use of multiple keys.

The added at field is left as a valid time value that is meaningless since this reference implementation does not define a time that the key was generated or added to the account at. Future versions of this implementation that stores signing keys per account can assign the date the signer was stored.