services/horizon/internal/actions: Use uint32 for ledger_id and transform error. (#2629)
### What

Extend `GetParams` to handle `schema` errors. 

### Why

`govalidator` is great for string query parameters, however, if the value represents something different to a string you have to add custom code to cast the value to their intended type, for example `string` to `uint`. 

This PR extends how we map query params to the destination struct so if there is a problem while using the query params to fill the params struct, we report the failing field and show the error, allowing us to add custom errors if necessary.