Update base branch for verify range job (#2668)
Recently we fixed a bug with how transaction fees were represented in postgres. Previously, we stored the fee as a 32 bit signed integer. But, it turns out that the range of fee values can reach the maximum unsigned 32 bit integer.

In the old code, when we ingested transactions those fee values would appear negative in the database. But, in the new code, the fee values are correctly represented as positive.

This PR updates the base branch to be the tag of the horizon 1.3.0 release which corresponds to the last release which was verified. When we use this base branch we will no longer get false alerts based on negative fee values.