exp/ingest/ledgerbackend: Captive Core: Online mode (#2694)
Adds online mode to `CaptiveStellarCore` backend:

* The `PrepareRange` method is now used to initialize an online mode
when `to` parameter is equal `0`. It also detects when the range is
already prepared: in such case it doesn't wait for the buffer to be
non-empty.
* `GetLedger` no longer starts Stellar-Core process. Users should always
call `PrepareRange` to define the mode (online/offline) and range they
want to fetch.
* `stellarCoreRunner` has a new `catchup` method. It runs `new-db`,
`catchup [X-1]/0` and `run` sequentially to make Stellar-Core start
streaming ledgers starting at ledger `X`.
* Updated `GetLatestLedgerSequence` to not use history archives. It now
only works when `PrepareRange` has been called and returns `lastLedger`
sequence in the offline mode and, currently, `nextLedger` or `nextLedger
- 1` in the online mode (see below).
* Updated ingestion state machine to use `PrepareRange` methods when
required.

This is still experimental code but works! We should fix fundamental
issues in this PR and merge it but work on new features and polish it in
another PRs.