Fix data race in root endpoint (#2745)
There is a data race in the root endpoint when the action accesses action.App.coreVersion and other properties from action.App. These properties are updated in the App.Tick() goroutine. This commit fixes the data race by wrapping these properties in a read write lock. Whenever the properties are updated the write lock is acquired. Similarly, whenever the properties are read the read lock is acquired.