support/keypairgen: add types and interface for keypair generation and mocking (#2762)
### What
Add type that wraps keypair generation, `keypairgen.Generator`, and interface, `keypairgen.Source`, that can be easily mocked.

The default source for the generator is the random source, i.e. `keypair.Random`.

A `keypairgentest.SliceSource` type is provided as well.

### Why
I'm adding functionality to recoverysigner that will have that service randomly generate a new signing key when an account registers. Writing tests for that logic is difficult because I can't predict what the signing key will be for an account that gets registered because if the application code calls keypair.Random directly there's no easy way to mock it. In the service I plan to mock the call to keypair.Random and the type and interface for doing so seems reusable.

The type and interface are heavily inspired by the `support/clock` package in this repository and follows the same patterns as that package, including a subpackage `keypairgentest` and a fixed value implementation of the source useful for testing.