Only allow non-decreasing calls to GetLedger (#2799)
Unlike DatabaseBackend instances, there is an implicit expectation that aCaptiveStellarCore instance should not be called on ledgers with a random access pattern. CaptiveStellarCore expects GetLedger() to be called on a sequence of non decreasing ledgers. It is possible for a CaptiveStellarCore to seek ahead to a future ledger in the stream. However, CaptiveStellarCore does not support rewinding back several ledgers in history.

This commit adds an explicit check to GetLedger() which returns an error if GetLedger() is called using non-decreasing ledgers.