exp/ingest/ledgerbackend: Handle Stellar-Core process exit gracefully (#2803)
This commit adds a code that handles Stellar-Core process exit
gracefully. Added `processExit` unbuffered channel to
`stellarCoreRunner` that receives the error in case Stellar-Core crashed
or was terminated with an error or nil error if it was shutdown
gracefully. The channel is later used in `PrepareRange` that returns an
error if the process was stopped while prepare was running and in
`sendLedgerMeta` that adds error to a meta channel that will be read by
one of the following `GetLedger` calls.