all: update go-xdr to 71a1e6d555f2 (#2814)
### What
Update stellar/go-xdr to 71a1e6d555f2.

### Why
There are no meaningful changes between the version of go-xdr we are
using and this new version. The changes between those versions were
all test fixes and linting fixes, most of which occurred when we
migrated the library to the stellar org on GitHub.

The reason we'd want to update the commit we're referencing is to remove
github.com/nullstyle/go-xdr from our dependency graph because it is only
being pulled into the graph because of tests in stellar/go-xdr
referencing it.