services/horizon/expingest: Update ingestion state machine to not block when captive backend is used (#2802)
This commit updates ingestion state machine to not block entire cluster
ingestion when captive core backend is not prepared. Specifically:

* In `buildState`, `resumeState` and `historyRangeState` added a check
that ensures a range is prepared. If not, it releases the distributed
ingestion lock and prepares the selected range.
* In `resumeState` it adds the code that fast-forwards the captive core
backend to the last ledger available in a backend buffer.
* Added `suggestedCheckpoint` to `startState` params that's set when
transitioning from `buildState`.

`CaptiveStellarCore` backend unlike `DatabaseBackend` does not allow
_random-access_ to ledger data. To work it first need to prepare a range
(using `PrepareRange` method) that can take a few minutes. Without
changes in this PR, the node that is actively ingesting and preparing a
range would block other nodes (potentially synced) from ingesting.

Another change compared to `DatabaseBackend` is that `GetLedger` method
must be called in a non-decreasing order of ledger sequences. It also
contains a small buffer of ledgers, `GetLatestLedgerSequence()` doesn't
return the latest ledger sequence closed by the network (in case of a
synced node) but the latest ledger available in a buffer (that can be
also the latest ledger sequence closed by the network but that's not
always true: ex. when catching up).

Because of the reasons above small changes to `resumeState` were
required. With the changes in this commit, it fast-forwards to the
latest ledger available in a backend to advance the buffer. Without it,
the node would stay in an infinite loop asking for a ledger that is in
front of the few ledgers in a buffer.

Finally, `suggestedCheckpoint` is required because a new checkpoint can
be created when `buildState` is still preparing a range. When this
happens, the code checking a bucket hash, would fail because
`GetLatestLedgerSequence()` would return the previous checkpoint (plus a
few ledgers in buffer) because it was used as a starting sequence in
`PrepareRange`.