exp/services/captivecore/internal: Implement RemoteCaptiveCore HTTP client (#2918)
This PR is a follow up to #2896 . In #2896 we added an http server wrapper around stellar core so that we could run a captive core instance on a separate machine from horizon.

This PR adds the corresponding HTTP client for the captive core server. The client satisfies the LedgerBackend interface and can be plugged into Horizon's ingestion system.