services/horizon: Integration tests (#2916)
This commit adds a simple integration tests framework for building
Horizon tests.

We need integration tests to check how Horizon services work together
with each other and Stellar-Core. This is especially important for
protocol upgrades where mocking is usually not enough to ensure that
everything works correctly and testing different ledger backends (like
captive core).

Additionally, a new CircleCI workspace was added to make the process as
easy as possible for developers. It is not started by default (`hold`)
so a developer can start it whenever it's worth checking the integration
tests.