all: update supported go version to 1.14 and 1.15 (#3044)
### What

Drop support for Go 1.13 from all modules, and update CI tests to run against Go 1.14 and 1.15.

### Why

The code in this repo is intended to be supported on the latest two releases of Go. Go 1.15 was released on August 11th, 2020, and we should be running tests against this version and no longer need to run tests against Go 1.13.

### Known limitations

The Go 1.15 docker images don't have a Debian Stretch variant and so I've used the Debian Buster variant and the artifact publishing is still using Go 1.14. This means we're continuing to publish artifacts built on Debian Stretch. If a Stretch variant of Go 1.15 docker images is released I'll update this, otherwise we have about 6 months until Go 1.16 is released and we'll need to start publishing with builds on Debian Buster. I've requested a Debian Stretch variant here: docker-library/golang#344.